+++ frontend/src/components/PublicPage.tsx
@@ -294,7 +294,10 @@ class PublicPage extends Component<IProps,IState> {
 														<a
 															href="#"
 															className="logout"
-															onClick={store.logOut}
+															onClick={() => {
+															store.logOut();
+															this.openLoginOrProfileModal();
+														  }}
 														>
 															<i className="fas fa-sign-out-alt" />&nbsp;
 														{i18next.t('LOGOUT')}