+++ frontend/src/components/modals/FavMixModal.tsx
@@ -41,7 +41,11 @@ class FavMixModal extends Component<IProps,IState> {
     			<div className="modal-dialog modal-md">
     				<div className="modal-content">
     					<ul className="modal-header">
-    						<label className="modal-title">{i18next.t('START_FAV_MIX')}</label>
+    						<h4 className="modal-title">{i18next.t('START_FAV_MIX')}</h4>
+							<button className="closeModal btn btn-action" 
+								onClick={() => ReactDOM.unmountComponentAtNode(document.getElementById('modal'))}>
+								<i className="fas fa-times"></i>
+							</button>
     					</ul>
     					<div className="modal-body">
     						<div className="automixUserlist">